testbot/web: Allow hiding the full log with a double-click.
This saves the user from having to search for the hard to find start
of the (build) log.
Special care is taken to not disable word-based selection when
double-clicking on the log text itself. To switch back to the summary
view one must therefore double-click to the right of the end-of-line.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>