testbot: Use tighter build timeouts.
There is no need to allow for a full rebuild if the patch only touches a
few modules. This is particularly significant for Wine builds since they
take longer.
Reducing the timeouts ensures the TestBot will not be blocked longer
than necessary should a task get stuck either due to an issue in the
patch, or to a network issue.
This also centralizes all build timeout calcultations in
GetBuildTimeout().
Finally this slightly changes a couple of timeouts in the scheduler and
activity modules because the $BuildTimeout and $ReconfigTimeout
parameters are going away.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>