testbot: Remove the HasConfigure patch impact.
Even if the patch provides an updated configure file the TestBot needs
to be able to recreate it in case another patch does not provide it.
Furthermore skipping the autoconf steps correctly is complex when a
patchset changes configure.ac in part 1 and provides configure in part
2, or provides both in part 1 and only some other configure.ac change
in part 2, etc.
So it's better to just run autoconf whenever configure.ac is patched.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>