testbot: Fix patchset handling in GetPatchImpacts().
Only the last part must be taken into account to determine if a rebuild
or testing is needed. But they are needed, the previous parts must be
taken into account for the rebuild, including the Autoconf and
MakeMakefile fields.
This also means the scope of the rebuild may be different from that of
the tests (we may need a full rebuild to run a single test). So
introduce RebuildRoot and RebuildModules which are similar to
PatchedRoot and PatchedModules but also take into account patches from
previous parts.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>