testbot: Add Missions support.
Missions allow the TestBot administrator to specify which tests a given
VM should perform for mailing list patches. For instance this can be
used to restrict a 64 bit Windows VM to 32 bit tests (useful when it
duplicates another VM with just a different locale).
Furthermore missions can get options. Currently only the 'nosubmit'
option is supported and it disables submitting the WineTest results
online.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>