testbot: Specify which test units to rerun with the 'test' mission option.
If set to 'all', then any non-test patch will cause a full WineTest run.
If set to 'module', then only the patched module's tests will be run.
If set to 'test' or unset, then the current behavior is preserved so
that only patched test units are run.
If set to 'build', then only verify that the patch builds.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>