testbot/SetWinLocale: Automate changing the Windows locales.
SetWinLocale leverages TestAgent and intl.cpl to automate changing the
Windows locales of the specified host: sorting and formatting, system
locale, display language, keyboard and location.
It can also create locale VMs for the TestBot, that is copies of a
base VM that allow the TestBot to run the tests in the specified locale.
Then it can recreate a specific locale VM snapshot from the base
snapshot, or refresh the snapshots of all the locale VMs of a given base
VM.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>