testbot/BackupVM: A script for backing up a Libvirt/QEmu VM.
The script supports backing up to either a borg repository, possibly
remote, the deduplication makes this ideal for backups; or tar files,
which is more suited for archival.
By default the script will refuse to perform the backup if it foundi
no snapshots. If that's expected pass the --no-snapshot option.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>