testbot/RestoreVM: A script to restore a Libirt/QEmu VM.
This allows restoring a VM from either a borg repository (potentially
remote), a set of compressed tar files or a directory created by the
--extract-only option.
RestoreVM automatically waits for the VM to not be in use before
restoring it, and minimises the time it takes to perform the final step.
RestoreVM --disk-only verifies that none of the Libvirt configuration
files changed and then only replaces the disks. This can be useful to
remove accumulated cruft from the qcow2 image files.
The VM can be renamed during restoration, in which case its uuid and
MAC address are also modified to avoid conflicts (but the guest
configuration such as hostname and IP address must be changed
manually).
With --extract-only RestoreVM will just extract the files to a
temporary directory and show the commands to use to manually put them
into place.
The --extract-only option can be combined with renaming the VM and the
directory will contain both sets of Libvirt configuration files which
allows comparing and verifying them.
In case of an error (such as if the VM is running and does not stop
within the imparted time) RestoreVM automatically switches to the
--extract-only mode. This allows continuing the restoration by passing
that directory as an argument once the issue has been resolved.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>