testbot/LibvirtTool: Fix FatalError() when the VM status is not as expected.
If the VM Status is not in the expected state then it no longer
belongs to this process. So don't even try to save it as that could
cause interference. Also don't notify the administrator about putting
it offline since that will not be happening.
However if the VM is merely in maintenance mode that means the tool is
being debugged so proceed normally except for the VM Status which must
not be modified to avoid inteference from the scheduler.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>