testbot/Engine: Ignore ChildPid & ChildDeadline for maintenance VMs.
For debugging it is useful to set a VM in maintenance mode and then
manually run scripts operating on that VM. When doing so ChildPid is
undefined since it is normally set by the Engine when starting the
script. But the scripts may modify other VM fields, including
ChildDeadline.
But when the Scheduler sees such a 'left-over' ChildDeadline field it
normally marks the VM as dirty and powers it off, thus interfering
with the script being tested.
So modify the Scheduler to ignore ChildDeadline for maintenance VMs
whether for cleanup or to compute when to reschedule next.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>