testbot: Separate running reconfig and winetest Wine VM tasks.
The reconfig task updates and rebuilds Wine and is a high priority
task because patches should always be applied on top of the latest
Wine version.
In contrast re-running WineTest normally does not impact the test
results much and thus should not get in the way of providing the
results to the developpers in a timely manner.
But priorities are set at the job level. So create a first high priority
job for the reconfig tasks, and create a separate low priority job if
and when these tasks complete successfully.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>