testbot/LibvirtTool: Don't put VMs in maintenance even after repeated errors.
The VM may have gone offline simply because of repeated revert
timeouts. This can happen if there is too much load on the host which
is a transient issue. But the administrator may not be available right
away to investigate and put the VM back online.

So warn the administrator, keep monitoring the offline VM but space out
the attempts to use the VM so the host can try running tasks on other
VMs.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>