testbot: Requeue the task if the VM is not ready for it.
In order to run a task a VM should be in the 'running' state and powered
on. If either check fails that's not the fault of the task and a second
attempt may succeed. So don't end the task with a boterror. Instead
requeue it and put the VM offline so its true state is evaluated (and
notify the administrator of course).
This implies only resetting the VM error count after successfully
running a task otherwise we could get into an infinite loop putting it
offline at the start of every task.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>