testbot/LibvirtTool: Create live snapshots as needed.
Live snapshots have to save the content of the VM's memory which takes
space in the disk image. They may also be specific to the host they
are running on (e.g. AMD vs. Intel CPU). And they may need to be
recreated after QEmu upgrades.
So VMs always have a powered off snapshot as reference and backups
made from the powered off state. With VMs that are configured for
autologin and autostart of TestAgentd this patch removes the need for
the administrator to manually recreate the live snapshots.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>