testbot/Submit: Rewrite the submit job page.
Preserve all parameters as the user moves from one step to the next and
back. This includes preserving and showing the selected extra VMs even
when the other extra VMs are hidden.
Also validate the parameters whenever the user moves forward so the page
never uses an invalid parameter.
Introduce a new VMRow structure to store information about which VMs are
compatible with the current job, which VMs to show, etc. This also
provides the basis for future extensions.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>