testbot/Submit: Let the user pick the Wine build and language to test.
The Wine VMs are duplicated in the list, one entry per build, so the
user can choose to test a patch or executable on one or all of the Wine
builds.
However the supported languages are presented in a drop-down list to
keep the list size reasonable. So only one can be tested per Wine
build.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Signed-off-by: Alexandre Julliard <julliard@winehq.org>